+++ doc/release_notes.rst
@@ -4,7 +4,9 @@ Release Notes
 Upcoming Version
 ----------------
 
-* The handling of `pandas` objects was improved. As `pandas` objects are fully aware of coordinates, their index and columns are now strictly taken into account. For example, when multiplying a `pandas.DataFrame` with variables, linopy will now check the alignment of indexes and re-index accordingly. Before, if the shapes of the axes were the same, the indexes of the Variable were forced in and the `pandas` indexes were effectively ignored. A warning was added for cases where users should expect changes of results with this version.
+* The handling of `pandas` objects was improved. As `pandas` objects are fully aware of coordinates, their index and columns are now strictly taken into account. For example, when multiplying a `pandas.DataFrame` with a variable, linopy now checks the alignment of indexes and reindexes accordingly. Previously, if the axis shapes were the same, the indexes of the variable were inserted and the `pandas` indexes were effectively ignored. A warning has been added for cases where users should expect changes to the results with this version. **Important**: This does not apply to overwriting the coordinates when one expression is added to another, e.g. "x + df" still overwrites the index of "df" when the dimensional shapes are aligned.
+
+